requirements.txt
requests
python-dotenv
paho-mqtt